refactor: move loading screen into screen-components
