+++ src/screen-components/loading-screen/index.ts
@@ -1,3 +1,4 @@
 export {LoadingScreenBoundary} from './LoadingScreenBoundary';
 export {LoadingScreen} from './LoadingScreen';
 export {useIsLoadingAppState} from './use-is-loading-app-state';
+export type {LoadingParams} from './types';
